trying to impl a control system

+++ Game6/inc/Sprite.hpp
@@ -7,8 +7,8 @@ class Sprite {
     public:
     // glm::vec2 gpu_tr, gpu_tl, gpu_br, gpu_bl;
     std::weak_ptr<SpriteSheet> sheet;
-    int x, y, w, h;
+    float x, y, w, h;
     Sprite();
-    Sprite(std::shared_ptr<SpriteSheet> &sheet, int x, int y, int w, int h);
+    Sprite(std::shared_ptr<SpriteSheet> &sheet, float x, float y, float w, float h);
     const Sprite &operator=(const Sprite &);
 };
